demos/demo_sift1M_vectodb.go v2

+++ vectodb.go
@@ -16,7 +16,7 @@ type VectoDB struct {
 	vdb_c  unsafe.Pointer
 }
 
-func NewVectoDB(workDir string, dim int64, metricType int, indexKey string, queryParams string) (vdb *VectoDB, err error) {
+func NewVectoDB(workDir string, dim int, metricType int, indexKey string, queryParams string) (vdb *VectoDB, err error) {
 	workDir_c := C.CString(workDir)
 	indexKey_c := C.CString(indexKey)
 	queryParams_c := C.CString(queryParams)
@@ -39,14 +39,14 @@ func (vdb *VectoDB) Destroy() (err error) {
  * Writer methods. There could be multiple writers.
  */
 
-func (vdb *VectoDB) ActivateIndex(index unsafe.Pointer, ntrain int64) (err error) {
+func (vdb *VectoDB) ActivateIndex(index unsafe.Pointer, ntrain int) (err error) {
 	vdb.rwlock.Lock()
 	defer vdb.rwlock.Unlock()
 	C.VectodbActivateIndex(vdb.vdb_c, index, C.long(ntrain))
 	return
 }
 
-func (vdb *VectoDB) AddWithIds(nb int64, xb []float32, xids []int64) (err error) {
+func (vdb *VectoDB) AddWithIds(nb int, xb []float32, xids []int64) (err error) {
 	vdb.rwlock.Lock()
 	defer vdb.rwlock.Unlock()
 	C.VectodbAddWithIds(vdb.vdb_c, C.long(nb), (*C.float)(&xb[0]), (*C.long)(&xids[0]))
@@ -57,25 +57,25 @@ func (vdb *VectoDB) AddWithIds(nb int64, xb []float32, xids []int64) (err error)
  * Reader methods. There could be multiple readers.
  */
 
-func (vdb *VectoDB) TryBuildIndex(exhaust_threshold int64) (index unsafe.Pointer, ntrain int64, err error) {
+func (vdb *VectoDB) TryBuildIndex(exhaust_threshold int) (index unsafe.Pointer, ntrain int, err error) {
 	vdb.rwlock.RLock()
 	defer vdb.rwlock.RUnlock()
 	var ntrain_c C.long
 	index = C.VectodbTryBuildIndex(vdb.vdb_c, C.long(exhaust_threshold), &ntrain_c)
-	ntrain = int64(ntrain_c)
+	ntrain = int(ntrain_c)
 	return
 }
 
-func (vdb *VectoDB) BuildIndex() (index unsafe.Pointer, ntrain int64, err error) {
+func (vdb *VectoDB) BuildIndex() (index unsafe.Pointer, ntrain int, err error) {
 	vdb.rwlock.RLock()
 	defer vdb.rwlock.RUnlock()
 	var ntrain_c C.long
 	index = C.VectodbBuildIndex(vdb.vdb_c, &ntrain_c)
-	ntrain = int64(ntrain_c)
+	ntrain = int(ntrain_c)
 	return
 }
 
-func (vdb *VectoDB) Search(nq int64, xq []float32, distances []float32, xids []int64) (err error) {
+func (vdb *VectoDB) Search(nq int, xq []float32, distances []float32, xids []int64) (err error) {
 	vdb.rwlock.RLock()
 	defer vdb.rwlock.RUnlock()
 	C.VectodbSearch(vdb.vdb_c, C.long(nq), (*C.float)(&xq[0]), (*C.float)(&distances[0]), (*C.long)(&xids[0]))
